requirements.txt
Flask-Migrate
Flask-WTF
flask
flask_sqlalchemy
gunicorn
jsonschema
pyyaml
requests